requirements.txt
amqp==2.5.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
bs4==0.0.1
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-bootstrap4==0.0.8
django-tables2==2.1.0
et-xmlfile==1.0.1
google-api-core==1.14.0
google-auth==1.6.3
google-cloud-vision==0.38.0
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
kombu==4.6.3
lxml==4.3.4
numpy==1.16.4
openpyxl==2.6.2
pandas==0.25.0
protobuf==3.9.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
